LPD-26707 Merge Tests
