--- modules/test/playwright/tests/marketplace-web/fixtures/dashboardPagesTest.ts
@@ -1,18 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import test from '@playwright/test';
-
-import {PublisherSolutionPage} from '../pages/publisherSolutionPage';
-
-const publisherDashboardPagesTest = test.extend<{
-	publisherSolutionPage: PublisherSolutionPage;
-}>({
-	publisherSolutionPage: async ({page}, use) => {
-		await use(new PublisherSolutionPage(page));
-	},
-});
-
-export {publisherDashboardPagesTest};
